+++ 9_sycl_of_hell/7_out_of_order.cpp
@@ -43,7 +43,7 @@ int main(int argc, char **argv) {
   Q.copy(A, A_host.data(), global_range, e).wait();
   sycl::free(A, Q);
 
-  for (size_t i = 0; i < global_range; i++)
+  for (int i = 0; i < global_range; i++)
     std::cout << "A_host[ " << i << " ] = " << A_host[i] << std::endl;
   return 0;
 }
